+++ lib/src/localdate.dart
@@ -137,7 +137,7 @@ class LocalDate implements Comparable<LocalDate> {
   /// [dateTime]: Value to convert into a Time Machine local date
   /// [calendar]: The calendar system to convert into, defaults to ISO calendar
   /// Returns: A new [LocalDate] with the same values as the specified `DateTime`.
-  factory LocalDate.fromDateTime(DateTime dateTime, [CalendarSystem calendar])
+  factory LocalDate.dateTime(DateTime dateTime, [CalendarSystem calendar])
   {
     int days = Platform.isWeb
         ? dateTime.millisecondsSinceEpoch ~/ TimeConstants.millisecondsPerDay
@@ -154,7 +154,7 @@ class LocalDate implements Comparable<LocalDate> {
   /// Returns: The date corresponding to the given week year / week of week year / day of week.
   ///
   /// see: https://en.wikipedia.org/wiki/ISO_week_date
-  factory LocalDate.fromWeekYearWeekAndDay(int weekYear, int weekOfWeekYear, DayOfWeek dayOfWeek)
+  factory LocalDate.weekYearWeekAndDay(int weekYear, int weekOfWeekYear, DayOfWeek dayOfWeek)
   => WeekYearRules.iso.getLocalDate(weekYear, weekOfWeekYear, dayOfWeek, CalendarSystem.iso);
 
   /// Returns the local date corresponding to a particular occurrence of a day-of-week
@@ -172,7 +172,7 @@ class LocalDate implements Comparable<LocalDate> {
   /// [dayOfWeek]: The day-of-week of the value to return.
   /// The date corresponding to the given year and month, on the given occurrence of the
   /// given day of week.
-  factory LocalDate.fromYearMonthWeekAndDay(int year, int month, int occurrence, DayOfWeek dayOfWeek)
+  factory LocalDate.yearMonthWeekAndDay(int year, int month, int occurrence, DayOfWeek dayOfWeek)
   {
     // This validates year and month as well as getting us a useful date.
     LocalDate startOfMonth = new LocalDate(year, month, 1);