+++ test/timezones/algiers_test.dart
@@ -31,7 +31,7 @@ void GetPeriod_BeforeLast()
 {
   Instant april1981 = new Instant.utc(1981, 4, 1, 0, 0);
   var actual = Algiers.getZoneInterval(april1981);
-  var expected = IZoneInterval.newZoneInterval("WET", new Instant.epochTime(microseconds: 341802000000000), new Instant.epochTime(microseconds: 357523200000000), Offset.zero, Offset.zero);
+  var expected = IZoneInterval.newZoneInterval("WET", new Instant(microseconds: 341802000000000), new Instant(microseconds: 357523200000000), Offset.zero, Offset.zero);
   expect(expected, actual);
 }
 
@@ -40,7 +40,7 @@ void GetPeriod_AfterLastTransition()
 {
   var may1981 = new ZonedDateTime.atStrictly(new LocalDateTime(1981, 5, 1, 0, 0, 1), DateTimeZone.utc).toInstant();
   var actual = Algiers.getZoneInterval(may1981);
-  var expected = IZoneInterval.newZoneInterval("CET", new Instant.epochTime(microseconds: 357523200000000), null, new Offset(TimeConstants.secondsPerHour), Offset.zero);
+  var expected = IZoneInterval.newZoneInterval("CET", new Instant(microseconds: 357523200000000), null, new Offset(TimeConstants.secondsPerHour), Offset.zero);
   expect(expected, actual);
 }